rename: 헤터 푸터 상위 폴더 삭제
